agent-plugin/requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.28.0
backports.zoneinfo==0.2.1
brotlipy==0.7.0
click==8.1.4
Deprecated==1.2.14
Django==4.2.3
django-apscheduler==0.6.2
django-cors-headers==4.2.0
django-cron==0.6.0
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-limiter==0.1.5
h11==0.14.0
httptools==0.6.0
importlib-resources==6.0.0
iso8601==1.1.0
Jinja2==3.1.2
kafka-python==2.0.2
limits==3.5.0
MarkupSafe==2.1.3
packaging==23.1
psutil==5.9.5
pydantic==2.0.2
pydantic_core==2.1.2
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
redis==4.6.0
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tortoise-orm==0.19.3
typing_extensions==4.7.1
tzlocal==5.0.1
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
zipp==3.16.2
